Do not reuse RovasDialog after it was destroyed
When all layers are deleted, the dialog gets destroy()-ed. If then new layers are added again, the previous instance must not be reused. Therefore now the dialog is newly created each time.

A TimeTrackingManager is introduced as a model with the dialog as its view. The model persists the state regardless if the dialog is destroyed.